k8s-django-app/employee/requirements/base.txt

Django==1.9.7
requests==2.18.4
python-dateutil==2.5.0
python-mimeparse==0.1.4
django-tastypie==0.14.0
psycopg2

# search engines
django-haystack==2.6.1
